Improve config variable types
- changes config variable `defaultValue` type to derive from `dataType` and
  `collectionType`
- updates config variable runtime values to derive from `dataType` and
  `collectionType`
