+++ packages/spectral/src/types/IntegrationDefinition.ts
@@ -1,69 +1,19 @@
 import {
-  DataSourceDefinition,
-  ConnectionDefinition,
   ActionPerformFunction,
   ActionPerformReturn,
   TriggerEventFunction,
   TriggerPerformFunction,
   Inputs,
   TriggerResult,
-  DataSourceType,
   TriggerPayload,
-  Connection,
-  JSONForm,
-  ObjectFieldMap,
-  ObjectSelection,
-  ConfigVarResultCollection,
-  Schedule,
-  CollectionDataSourceType,
-  ComponentManifest,
+  ComponentRegistry,
+  TriggerReference,
+  ConfigVars,
+  ConfigPages,
+  UserLevelConfigPages,
+  ValueExpression,
+  ConfigVarExpression,
 } from ".";
-import { Prettify, UnionToIntersection } from "./utils";
-
-/**
- * Root ComponentRegistry type exposed for augmentation.
- *
- * The expected interface when augmenting is:
- *
- * ```ts
- * interface IntegrationDefinitionComponentRegistry {
- *   [key: string]: ComponentManifest
- * }
- * ```
- *
- */
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface IntegrationDefinitionComponentRegistry {}
-
-/**
- * Root ConfigPages type exposed for augmentation.
- *
- * The expected interface when augmenting is:
- *
- * ```ts
- * interface IntegrationDefinitionConfigPages {
- *   [key: string]: ConfigPage
- * }
- * ```
- *
- */
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface IntegrationDefinitionConfigPages {}
-
-/**
- * Root UserLevelConfigPages type exposed for augmentation.
- *
- * The expected interface when augmenting is:
- *
- * ```ts
- * interface IntegrationDefinitionUserLevelConfigPages {
- *   [key: string]: ConfigPage
- * }
- * ```
- *
- */
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface IntegrationDefinitionUserLevelConfigPages {}
 
 /** Defines attributes of a Code-Native Integration. */
 export type IntegrationDefinition = {
@@ -153,328 +103,6 @@ export interface Flow<TTriggerPayload extends TriggerPayload = TriggerPayload> {
   >;
 }
 
-export type PermissionAndVisibilityType =
-  | "customer"
-  | "embedded"
-  | "organization";
-
-/** Common attribute shared by all types of Config Vars. */
-type BaseConfigVar = {
-  /** A unique, unchanging value that is used to maintain identity for the Config Var even if the key changes. */
-  stableKey: string;
-  /** Optional description for this Config Var. */
-  description?: string;
-  /**
-   * Optional value that sets the permission and visibility of the Config Var. @default "customer"
-   *
-   * "customer" - Customers can view and edit the Config Var.
-   * "embedded" - Customers cannot view or update the Config Var as the value will be set programmatically.
-   * "organization" - Customers cannot view or update the Config Var as it will always have a default value or be set by the organization.
-   */
-  permissionAndVisibilityType?: PermissionAndVisibilityType;
-  /** Optional value that specifies whether this Config Var is visible to an Organization deployer. @default true */
-  visibleToOrgDeployer?: boolean;
-  /** Optional default value for the Config Var. */
-  defaultValue?: string;
-  /** Optional list of picklist values if the Config Var is a multi-choice selection value. */
-  pickList?: string[];
-  /** Optional value to specify the type of collection if the Config Var is multi-value. */
-  collectionType?: CollectionType;
-};
-
-/** Defines attributes of a standard Config Var. */
-export type StandardConfigVar = BaseConfigVar & {
-  /** The data type of the Config Var. */
-  dataType: Exclude<ConfigVarDataType, "schedule" | "code">;
-};
-
-export type CodeConfigVar = BaseConfigVar & {
-  /** The data type of the Config Var. */
-  dataType: "code";
-  /** Optional default value for the Config Var. */
-  defaultValue?: string;
-  /** Optional value to specify the type of language if the Config Var is a code value. */
-  codeLanguage?: CodeLanguageType;
-};
-
-export type ScheduleConfigVar = Omit<BaseConfigVar, "collectionType"> & {
-  /** The data type of the Config Var. */
-  dataType: "schedule";
-  /** Optional default value for the Config Var. */
-  defaultValue?: string;
-  /** Optional timezone for the schedule. */
-  timeZone?: string;
-};
-
-export type ComponentRegistry =
-  keyof IntegrationDefinitionComponentRegistry extends never
-    ? {
-        [key: string]: ComponentManifest;
-      }
-    : UnionToIntersection<
-        keyof IntegrationDefinitionComponentRegistry extends infer TComponentKey
-          ? TComponentKey extends keyof IntegrationDefinitionComponentRegistry
-            ? {
-                [Key in TComponentKey]: IntegrationDefinitionComponentRegistry[TComponentKey];
-              }
-            : never
-          : never
-      >;
-
-interface ConnectionInputPermissionAndVisibility {
-  /**
-   * Optional value that sets the permission and visibility of the Config Var. @default "customer"
-   *
-   * "customer" - Customers can view and edit the Config Var.
-   * "embedded" - Customers cannot view or update the Config Var as the value will be set programmatically.
-   * "organization" - Customers cannot view or update the Config Var as it will always have a default value or be set by the organization.
-   */
-  permissionAndVisibilityType?: PermissionAndVisibilityType;
-  /** Optional value that specifies whether this Config Var is visible to an Organization deployer. @default true */
-  visibleToOrgDeployer?: boolean;
-}
-
-type ComponentReferenceType = Extract<
-  keyof ComponentManifest,
-  "actions" | "triggers" | "dataSources" | "connections"
->;
-
-type ComponentReferenceTypeValueMap<
-  TValue,
-  TMap extends Record<ComponentReferenceType, unknown> = {
-    actions: ValueExpression<TValue>;
-    triggers: ValueExpression<TValue> | ConfigVarExpression;
-    dataSources: ValueExpression<TValue> | ConfigVarExpression;
-    connections: (ValueExpression<TValue> | ConfigVarExpression) &
-      ConnectionInputPermissionAndVisibility;
-  }
-> = TMap;
-
-export type ComponentReference<
-  TComponentReference extends {
-    component: string;
-    key: string;
-    values?: {
-      [key: string]: ValueExpression | ConfigVarExpression;
-    };
-    template?: string;
-  } = {
-    component: string;
-    key: string;
-    values?: {
-      [key: string]: ValueExpression | ConfigVarExpression;
-    };
-    template?: string;
-  }
-> = TComponentReference;
-
-export const isComponentReference = (ref: unknown): ref is ComponentReference =>
-  typeof ref === "object" && ref !== null && "key" in ref && "component" in ref;
-
-type ComponentReferencesByType = UnionToIntersection<
-  ComponentReferenceType extends infer TComponentReferenceType
-    ? TComponentReferenceType extends Extract<
-        keyof ComponentManifest,
-        "actions" | "triggers" | "dataSources" | "connections"
-      >
-      ? {
-          [Key in TComponentReferenceType]: keyof ComponentRegistry extends infer TComponentKey
-            ? TComponentKey extends keyof ComponentRegistry
-              ? TComponentKey extends string
-                ? TComponentReferenceType extends keyof ComponentRegistry[TComponentKey]
-                  ? keyof ComponentRegistry[TComponentKey][TComponentReferenceType] extends infer TComponentPropertyKey
-                    ? TComponentPropertyKey extends keyof ComponentRegistry[TComponentKey][TComponentReferenceType]
-                      ? TComponentPropertyKey extends string
-                        ? "perform" extends keyof ComponentRegistry[TComponentKey][TComponentReferenceType][TComponentPropertyKey]
-                          ? ComponentRegistry[TComponentKey][TComponentReferenceType][TComponentPropertyKey]["perform"] extends (
-                              ...args: any[]
-                            ) => any
-                            ? Parameters<
-                                ComponentRegistry[TComponentKey][TComponentReferenceType][TComponentPropertyKey]["perform"]
-                              >[0] extends infer TInputs
-                              ? ComponentReference<{
-                                  component: TComponentKey;
-                                  key: TComponentPropertyKey;
-                                  values: {
-                                    [Key in keyof TInputs]: ComponentReferenceTypeValueMap<
-                                      TInputs[Key]
-                                    >[TComponentReferenceType];
-                                  };
-                                }>
-                              : never
-                            : never
-                          : never
-                        : never
-                      : never
-                    : never
-                  : never
-                : never
-              : never
-            : never;
-        }
-      : never
-    : never
->;
-
-export type TriggerReference = ComponentReferencesByType["triggers"];
-
-export type ActionReference = ComponentReferencesByType["actions"];
-
-export type DataSourceReference = ComponentReferencesByType["dataSources"];
-
-export type ConnectionReference = ComponentReferencesByType["connections"];
-
-// Handle some data source types not supporting collections.
-type BaseDataSourceConfigVar =
-  | ({
-      dataSourceType: CollectionDataSourceType;
-    } & BaseConfigVar)
-  | ({
-      dataSourceType: Exclude<DataSourceType, CollectionDataSourceType>;
-    } & Omit<BaseConfigVar, "collectionType">);
-type DataSourceDefinitionConfigVar = BaseDataSourceConfigVar &
-  Omit<
-    DataSourceDefinition<Inputs, ConfigVarResultCollection, DataSourceType>,
-    | "display"
-    | "inputs"
-    | "examplePayload"
-    | "dataSourceType"
-    | "detailDataSource"
-  >;
-export type DataSourceReferenceConfigVar = BaseDataSourceConfigVar & {
-  dataSource: DataSourceReference;
-};
-
-/** Defines attributes of a data source Config Var. */
-export type DataSourceConfigVar =
-  | DataSourceDefinitionConfigVar
-  | DataSourceReferenceConfigVar;
-
-type BaseConnectionConfigVar = Omit<BaseConfigVar, "collectionType"> & {
-  dataType: "connection";
-};
-
-type ConnectionDefinitionConfigVar = BaseConnectionConfigVar &
-  Omit<ConnectionDefinition, "label" | "comments" | "key">;
-
-type ConnectionReferenceConfigVar = BaseConnectionConfigVar & {
-  connection: ConnectionReference & {
-    template?: string;
-  };
-};
-
-/** Defines attributes of a Config Var that represents a Connection. */
-export type ConnectionConfigVar =
-  | ConnectionDefinitionConfigVar
-  | ConnectionReferenceConfigVar;
-
-export type ConfigVar =
-  | StandardConfigVar
-  | CodeConfigVar
-  | ScheduleConfigVar
-  | DataSourceConfigVar
-  | ConnectionConfigVar;
-
-export const isCodeConfigVar = (cv: ConfigVar): cv is CodeConfigVar =>
-  "dataType" in cv && cv.dataType === "code";
-
-export const isScheduleConfigVar = (cv: ConfigVar): cv is ScheduleConfigVar =>
-  "dataType" in cv && cv.dataType === "schedule";
-
-export const isDataSourceDefinitionConfigVar = (
-  cv: ConfigVar
-): cv is DataSourceDefinitionConfigVar =>
-  "dataSourceType" in cv && "perform" in cv && typeof cv.perform === "function";
-
-export const isDataSourceReferenceConfigVar = (
-  cv: ConfigVar
-): cv is DataSourceReferenceConfigVar =>
-  "dataSourceType" in cv &&
-  "dataSource" in cv &&
-  isComponentReference(cv.dataSource);
-
-export const isConnectionDefinitionConfigVar = (
-  cv: ConfigVar
-): cv is ConnectionDefinitionConfigVar =>
-  "dataType" in cv && cv.dataType === "connection" && "inputs" in cv;
-
-export const isConnectionReferenceConfigVar = (
-  cv: ConfigVar
-): cv is ConnectionReferenceConfigVar =>
-  "connection" in cv && isComponentReference(cv.connection);
-
-export type ConfigPageElement = string | ConfigVar;
-
-type CreateConfigPages<TIntegrationDefinitionConfigPages> =
-  keyof TIntegrationDefinitionConfigPages extends never
-    ? { [key: string]: ConfigPage }
-    : UnionToIntersection<
-        keyof TIntegrationDefinitionConfigPages extends infer TPageName
-          ? TPageName extends keyof TIntegrationDefinitionConfigPages
-            ? TIntegrationDefinitionConfigPages[TPageName] extends ConfigPage
-              ? {
-                  [Key in TPageName]: TIntegrationDefinitionConfigPages[TPageName];
-                }
-              : never
-            : never
-          : never
-      >;
-
-export type ConfigPages = CreateConfigPages<IntegrationDefinitionConfigPages>;
-export type UserLevelConfigPages =
-  CreateConfigPages<IntegrationDefinitionUserLevelConfigPages>;
-
-type ExtractConfigVars<TConfigPages extends { [key: string]: ConfigPage }> =
-  keyof TConfigPages extends infer TPageName
-    ? TPageName extends keyof TConfigPages
-      ? TConfigPages[TPageName] extends infer TConfigPage
-        ? TConfigPage extends ConfigPage
-          ? {
-              [Key in keyof TConfigPage["elements"] as Key extends string
-                ? TConfigPage["elements"][Key] extends ConfigVar
-                  ? Key
-                  : never
-                : never]: ElementToRuntimeType<TConfigPage["elements"][Key]>;
-            }
-          : never
-        : never
-      : never
-    : never;
-
-export type ConfigVars = Prettify<
-  UnionToIntersection<
-    ExtractConfigVars<ConfigPages> | ExtractConfigVars<UserLevelConfigPages>
-  >
->;
-
-export type ToDataSourceRuntimeType<TType extends DataSourceType> =
-  TType extends "jsonForm"
-    ? JSONForm
-    : TType extends "objectSelection"
-    ? ObjectSelection
-    : TType extends "objectFieldMap"
-    ? ObjectFieldMap
-    : string;
-
-export type ElementToRuntimeType<TElement extends ConfigPageElement> =
-  TElement extends ConnectionConfigVar
-    ? Connection
-    : TElement extends DataSourceConfigVar
-    ? ToDataSourceRuntimeType<TElement["dataSourceType"]>
-    : TElement extends ScheduleConfigVar
-    ? Schedule
-    : TElement extends StandardConfigVar
-    ? string
-    : never;
-
-/** Defines attributes of a Config Wizard Page used when deploying an Instance of an Integration. */
-export interface ConfigPage {
-  /** Elements included on this Config Page. */
-  elements: Record<string, ConfigPageElement>;
-  /** Specifies an optional tagline for this Config Page. */
-  tagline?: string;
-}
-
 /** Defines attributes of a Preprocess Flow Configuration used by a Flow of an Integration. */
 export type PreprocessFlowConfig = {
   /** Name of the field in the data payload returned by the Preprocess Flow to use for a Flow Name. */
@@ -526,28 +154,3 @@ export type EndpointSecurityType =
 
 /** Choices of Step Error Handlers that define the behavior when a step error occurs. */
 export type StepErrorHandlerType = "fail" | "ignore" | "retry";
-
-/** Supported data types for Config Vars. */
-export type ConfigVarDataType =
-  | "string"
-  | "date"
-  | "timestamp"
-  | "picklist"
-  | "schedule"
-  | "code"
-  | "boolean"
-  | "number";
-
-/** Choices of programming languages that may be used for Config Var code values. */
-export type CodeLanguageType = "json" | "xml" | "html";
-
-/** Choices of collection types for multi-value Config Vars. */
-export type CollectionType = "valuelist" | "keyvaluelist";
-
-/** Choices of component reference types. */
-export type ComponentSelectorType = "trigger" | "connection" | "dataSource";
-
-export type ConfigVarExpression = { configVar: string };
-export type ValueExpression<TValueType = unknown> = {
-  value: TValueType;
-};
